+++ app/src/main/java/com/mookseong/subwaysupport/data/entity/subway/DriveInfo.kt
@@ -1,4 +1,4 @@
-package com.mookseong.subwaysupport.data.subway
+package com.mookseong.subwaysupport.data.entity.subway
 
 data class DriveInfo(
     val laneID : String,
