+++ app/src/main/java/com/mookseong/subwaysupport/data/entity/subway/DriveInfoSet.kt
@@ -1,4 +1,4 @@
-package com.mookseong.subwaysupport.data.subway
+package com.mookseong.subwaysupport.data.entity.subway
 
 data class DriveInfoSet(
     val driveInfo : ArrayList<DriveInfo>
