+++ app/src/main/java/com/mookseong/subwaysupport/ui/startUp/StartupFragment.kt
@@ -2,7 +2,7 @@ package com.mookseong.subwaysupport.ui.startUp
 
 import androidx.lifecycle.ViewModelProvider
 import com.mookseong.subwaysupport.R
-import com.mookseong.subwaysupport.base.BaseFragment
+import com.mookseong.subwaysupport.ui.base.BaseFragment
 import com.mookseong.subwaysupport.databinding.FragmentStartupBinding
 
 class StartupFragment : BaseFragment<FragmentStartupBinding, StartupViewModel>(R.layout.fragment_startup) {
